Ship connection system working.
Really need to clean up the data containers in use; having a separate ship type similar to a Character, and then having a Location container also called Ship is somewhat confusing. Considering stripping the mechanics out of the map to handle Ship stuff and stuffing it into the like-a-character objects.